+++ kodata/shell/upgrade.sh
@@ -96,6 +96,9 @@ kubectl apply -f $KO_DATA_PATH/yaml/longhorn/default-data-locality.yaml || echo 
 
 
 echo "higress config"
+echo "安装 Gateway API CRD"
+kubectl apply -f https://github.com/kubernetes-sigs/gateway-api/releases/download/v1.4.0/experimental-install.yaml
+kubectl apply -f "$KO_DATA_PATH/yaml/higress-gateway.yaml"
 # higress 可能未启动成功 导致crd未创建 job设置重试3次
 kubectl apply -f $KO_DATA_PATH/yaml/higress-compressor.yaml --server-side